migrate classes
