+++ pkg/pgdb/models/organizationStorage.go
@@ -1,0 +1,6 @@
+package models
+
+type OrganizationStorage struct {
+	OrganizationId int64 `json:"organization_id"`
+	Size           int64 `json:"size"`
+}
