+++ pkg/pgdb/models/organizationUserTable.go
@@ -1,9 +1,6 @@
-package dbTable
+package models
 
 import (
-	"database/sql"
-	"fmt"
-	"github.com/pennsieve/pennsieve-go-core/pkg/core"
 	"time"
 )
 
@@ -46,29 +43,4 @@ type OrganizationUser struct {
 	DbPermission   DbPermission `json:"permission_bit"`
 	CreatedAt      time.Time    `json:"created_at"`
 	UpdatedAt      time.Time    `json:"updated_at"`
-}
-
-func (o *OrganizationUser) GetByUserId(db core.PostgresAPI, id int64) (*OrganizationUser, error) {
-
-	queryStr := "SELECT organization_id, user_id, permission_bit, created_at, updated_at " +
-		"FROM pennsieve.organization_user WHERE user_id=$1;"
-
-	var orgUser OrganizationUser
-	row := db.QueryRow(queryStr, id)
-	err := row.Scan(
-		&orgUser.OrganizationId,
-		&orgUser.UserId,
-		&orgUser.DbPermission,
-		&orgUser.CreatedAt,
-		&orgUser.UpdatedAt)
-
-	switch err {
-	case sql.ErrNoRows:
-		fmt.Println("No rows were returned!")
-		return nil, err
-	case nil:
-		return &orgUser, nil
-	default:
-		panic(err)
-	}
 }
