+++ pkg/pgdb/models/packageStorage.go
@@ -1,0 +1,6 @@
+package models
+
+type PackageStorage struct {
+	PackageId int64 `json:"package_id"`
+	Size      int64 `json:"size"`
+}
